Backend/requirements.txt
fastapi == 0.112.0
uvicorn == 0.32.1
python-dotenv == 1.0.1
pydantic == 2.10.1
motor == 3.6.0
beanie == 1.27.0
pytest == 8.3.3
pytest-asyncio == 0.24.0
httpx == 0.27.2